refactor(keyboard): refactor global shortcuts to use KeyboardProvider context
- Wrapped app layout with KeyboardProvider for centralized keyboard context.
- Replaced manual global keydown event listener in useGlobalShortcuts with useKeyboardContext hook.
- Simplified shortcut handling with declarative shortcut definitions and handlers.
- Improved management of next issue recommendation shortcut to prevent multiple executions.
- Maintained existing shortcut functionality with cleaner and more maintainable code.
