+++ hooks/use-global-shortcuts.tsx
@@ -1,11 +1,12 @@
-"use client"
+'use client'
 
-import * as React from "react"
-import { useRouter } from "next/navigation"
-import { useCommandPalette } from "./use-command-palette"
-import { createClient } from "@/lib/supabase/client"
-import { emitIssueStatusUpdate } from "@/lib/events/issue-events"
-import { useToast } from "@/components/ui/use-toast"
+import * as React from 'react'
+import { useRouter } from 'next/navigation'
+import { useCommandPalette } from '@/hooks/use-command-palette'
+import { createClient } from '@/lib/supabase/client'
+import { emitIssueStatusUpdate } from '@/lib/events/issue-events'
+import { useKeyboardContext, KeyboardPriority } from '@/lib/keyboard'
+import { useToast } from '@/components/ui/use-toast'
 
 interface GlobalShortcutsProps {
   workspaceSlug: string
@@ -20,14 +21,19 @@ interface GlobalShortcutsProps {
   onIssueStatusChange?: (newStatus: string) => void
 }
 
-export function useGlobalShortcuts({ workspaceSlug, onCreateIssue, onShowHelp, onToggleViewMode, currentIssue, onIssueStatusChange }: GlobalShortcutsProps) {
+export function useGlobalShortcuts({ 
+  workspaceSlug, 
+  onCreateIssue, 
+  onShowHelp, 
+  onToggleViewMode, 
+  currentIssue, 
+  onIssueStatusChange 
+}: GlobalShortcutsProps) {
   const router = useRouter()
-  const { setIsOpen: setCommandPaletteOpen } = useCommandPalette()
-  const { toast } = useToast()
-  const [keySequence, setKeySequence] = React.useState<string[]>([])
-  const sequenceTimeoutRef = React.useRef<NodeJS.Timeout | undefined>(undefined)
   const nextIssueTimeoutRef = React.useRef<NodeJS.Timeout | undefined>(undefined)
   const isExecutingNextIssue = React.useRef(false)
+  const { setIsOpen: setCommandPaletteOpen } = useCommandPalette()
+  const { toast } = useToast()
 
   const handleStatusChange = async (newStatus: string) => {
     console.log('Attempting status change:', { currentIssue, newStatus })
@@ -43,7 +49,6 @@ export function useGlobalShortcuts({ workspaceSlug, onCreateIssue, onShowHelp, o
     if (!error) {
       onIssueStatusChange(newStatus)
       console.log('Status changed successfully to:', newStatus)
-      // Emit event for other components
       emitIssueStatusUpdate(currentIssue.id, newStatus)
       toast({
         title: "Status updated",
@@ -59,180 +64,167 @@ export function useGlobalShortcuts({ workspaceSlug, onCreateIssue, onShowHelp, o
     }
   }
 
-  React.useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      // Ignore if user is typing in an input, textarea, or contenteditable
-      const target = e.target as HTMLElement
-      if (
-        target.tagName === "INPUT" ||
-        target.tagName === "TEXTAREA" ||
-        target.contentEditable === "true" ||
-        target.closest('[role="dialog"]') // Ignore if a dialog is open
-      ) {
-        return
-      }
-
-      // Handle Cmd/Ctrl + K (already handled by command palette hook, but prevent default)
-      if ((e.metaKey || e.ctrlKey) && e.key === "k") {
-        e.preventDefault()
-        return
-      }
-
-      // Handle Cmd/Ctrl + B for toggling view mode
-      if ((e.metaKey || e.ctrlKey) && e.key === "b") {
-        e.preventDefault()
-        onToggleViewMode?.()
-        return
-      }
-
-      // Handle Cmd/Ctrl + N for Next Issue recommendation
-      if ((e.metaKey || e.ctrlKey) && e.key === "n") {
-        e.preventDefault()
-        
-        // Prevent multiple simultaneous executions
-        if (isExecutingNextIssue.current) {
-          return
-        }
-        
-        // Clear any existing timeout
-        if (nextIssueTimeoutRef.current) {
-          clearTimeout(nextIssueTimeoutRef.current)
-        }
-        
-        isExecutingNextIssue.current = true
-        
-        // Trigger command palette with Next Issue command
-        setCommandPaletteOpen(true)
-        
-        // Use a ref to store the timeout so we can clean it up
-        nextIssueTimeoutRef.current = setTimeout(() => {
-          const nextIssueButton = document.querySelector('[data-command-id="next-issue"]') as HTMLElement
-          if (nextIssueButton) {
-            nextIssueButton.click()
+  // Register global keyboard shortcuts
+  useKeyboardContext({
+    id: 'global-shortcuts',
+    priority: KeyboardPriority.GLOBAL,
+    enabled: true,
+    shortcuts: {
+      // Command shortcuts
+      'cmd+k': {
+        handler: () => {
+          // Command palette hook handles this, just prevent default
+          return true
+        },
+        description: 'Open command palette',
+      },
+      'ctrl+k': {
+        handler: () => {
+          // Command palette hook handles this, just prevent default
+          return true
+        },
+        description: 'Open command palette',
+      },
+      'cmd+b': {
+        handler: () => {
+          onToggleViewMode?.()
+          return true
+        },
+        description: 'Toggle view mode',
+      },
+      'ctrl+b': {
+        handler: () => {
+          onToggleViewMode?.()
+          return true
+        },
+        description: 'Toggle view mode',
+      },
+      'cmd+n': {
+        handler: () => {
+          // Prevent multiple simultaneous executions
+          if (isExecutingNextIssue.current) {
+            return true
           }
-          isExecutingNextIssue.current = false
-          nextIssueTimeoutRef.current = undefined
-        }, 100)
-        return
-      }
-
-      // Forward slash is now handled by search bar in workspace-content.tsx
-      // Commented out to allow search bar toggle functionality
-      // if (e.key === "/") {
-      //   e.preventDefault()
-      //   setCommandPaletteOpen(true)
-      //   return
-      // }
-
-      // Handle question mark for help
-      if (e.key === "?" && e.shiftKey) {
-        e.preventDefault()
-        onShowHelp?.()
-        return
-      }
-
-      // Handle single key shortcuts
-      const key = e.key.toLowerCase()
-
-      // Check for sequential shortcuts
-      if (keySequence.length > 0) {
-        const sequence = [...keySequence, key].join("")
-        console.log('Key sequence detected:', sequence)
-        
-        // G then I - Go to Issues
-        if (sequence === "gi") {
-          e.preventDefault()
+          
+          // Clear any existing timeout
+          if (nextIssueTimeoutRef.current) {
+            clearTimeout(nextIssueTimeoutRef.current)
+          }
+          
+          isExecutingNextIssue.current = true
+          setCommandPaletteOpen(true)
+          
+          nextIssueTimeoutRef.current = setTimeout(() => {
+            const nextIssueButton = document.querySelector('[data-command-id="next-issue"]') as HTMLElement
+            if (nextIssueButton) {
+              nextIssueButton.click()
+            }
+            isExecutingNextIssue.current = false
+            nextIssueTimeoutRef.current = undefined
+          }, 100)
+          return true
+        },
+        description: 'Next issue recommendation',
+      },
+      'ctrl+n': {
+        handler: () => {
+          // Prevent multiple simultaneous executions
+          if (isExecutingNextIssue.current) {
+            return true
+          }
+          
+          // Clear any existing timeout
+          if (nextIssueTimeoutRef.current) {
+            clearTimeout(nextIssueTimeoutRef.current)
+          }
+          
+          isExecutingNextIssue.current = true
+          setCommandPaletteOpen(true)
+          
+          nextIssueTimeoutRef.current = setTimeout(() => {
+            const nextIssueButton = document.querySelector('[data-command-id="next-issue"]') as HTMLElement
+            if (nextIssueButton) {
+              nextIssueButton.click()
+            }
+            isExecutingNextIssue.current = false
+            nextIssueTimeoutRef.current = undefined
+          }, 100)
+          return true
+        },
+        description: 'Next issue recommendation',
+      },
+      
+      // Single key shortcuts
+      'shift+?': {
+        handler: () => {
+          onShowHelp?.()
+          return true
+        },
+        description: 'Show keyboard shortcuts',
+      },
+      'c': {
+        handler: () => {
+          onCreateIssue?.()
+          return true
+        },
+        description: 'Create new issue',
+      },
+      
+      // Sequential shortcuts - Navigation
+      'gi': {
+        handler: () => {
           router.push(`/${workspaceSlug}`)
-          setKeySequence([])
-          return
-        }
-        
-        // G then N - Go to Inbox
-        if (sequence === "gn") {
-          e.preventDefault()
+          return true
+        },
+        description: 'Go to issues',
+      },
+      'gn': {
+        handler: () => {
           router.push(`/${workspaceSlug}/inbox`)
-          setKeySequence([])
-          return
-        }
-
-        // Status change shortcuts (only on issue details page)
-        if (currentIssue && onIssueStatusChange) {
-          // S then T - Change to Todo
-          if (sequence === "st") {
-            e.preventDefault()
-            handleStatusChange("todo")
-            setKeySequence([])
-            return
-          }
-          
-          // S then P - Change to In Progress
-          if (sequence === "sp") {
-            e.preventDefault()
-            handleStatusChange("in_progress")
-            setKeySequence([])
-            return
-          }
-          
-          // S then R - Change to In Review
-          if (sequence === "sr") {
-            e.preventDefault()
-            handleStatusChange("in_review")
-            setKeySequence([])
-            return
-          }
-          
-          // S then D - Change to Done
-          if (sequence === "sd") {
-            e.preventDefault()
-            handleStatusChange("done")
-            setKeySequence([])
-            return
-          }
-        }
-
-        // Clear sequence if no match
-        setKeySequence([])
-        clearTimeout(sequenceTimeoutRef.current)
-      }
-
-      // Start a sequence with G or S
-      if (key === "g" || (key === "s" && currentIssue)) {
-        e.preventDefault()
-        setKeySequence([key])
-        
-        // Clear sequence after 1 second if no follow-up
-        clearTimeout(sequenceTimeoutRef.current)
-        sequenceTimeoutRef.current = setTimeout(() => {
-          setKeySequence([])
-        }, 1000)
-        return
-      }
-
-      // Single key shortcuts (only if no modifier keys are pressed)
-      if (!e.ctrlKey && !e.metaKey && !e.altKey && !e.shiftKey) {
-        switch (key) {
-          case "c":
-            e.preventDefault()
-            onCreateIssue?.()
-            break
-          case "f":
-            e.preventDefault()
-            // TODO: Implement filter
-            console.log("Filter shortcut pressed")
-            break
-        }
-      }
-    }
-
-    window.addEventListener("keydown", handleKeyDown)
+          return true
+        },
+        description: 'Go to inbox',
+      },
+      
+      // Sequential shortcuts - Status changes
+      'st': {
+        handler: () => {
+          handleStatusChange('todo')
+          return true
+        },
+        description: 'Set status to Todo',
+      },
+      'sp': {
+        handler: () => {
+          handleStatusChange('in_progress')
+          return true
+        },
+        description: 'Set status to In Progress',
+      },
+      'sr': {
+        handler: () => {
+          handleStatusChange('in_review')
+          return true
+        },
+        description: 'Set status to In Review',
+      },
+      'sd': {
+        handler: () => {
+          handleStatusChange('done')
+          return true
+        },
+        description: 'Set status to Done',
+      },
+    },
+    deps: [workspaceSlug, onCreateIssue, onShowHelp, onToggleViewMode, currentIssue, onIssueStatusChange],
+  })
+  
+  // Cleanup on unmount
+  React.useEffect(() => {
     return () => {
-      window.removeEventListener("keydown", handleKeyDown)
-      clearTimeout(sequenceTimeoutRef.current)
       if (nextIssueTimeoutRef.current) {
         clearTimeout(nextIssueTimeoutRef.current)
       }
     }
-  }, [workspaceSlug, router, onCreateIssue, onShowHelp, onToggleViewMode, setCommandPaletteOpen, keySequence, currentIssue, handleStatusChange])
-
-  return { keySequence }
+  }, [])
 }
